default not push "meta" branch on GitHub
This build script is now stable enough, and has error detection mechanism,
there is no need to keep pushing meta data/articats on GitHub as wasting the resource.

We can still enable it via set pushMetaOnGitHub=true in config.

+++ update-website.sh
@@ -124,7 +124,9 @@ elif [ "$updateRepo" = true ]; then
     msg="Now push and deploy website only, no need to deploy api due to meta data no update"
     output Success "$msg" gitter
     git push heroku meta:master -f || output Warn "Error" gitter
-    git push origin meta -f || output Warn "Error" gitter
+    if [ "$pushMetaOnGitHub" = true ]; then
+        git push origin meta -f || output Warn "Error" gitter
+    fi
 else
     msg="Didn't update anything, no need to push or deploy."
     output Info "$msg" gitter
